refactor NewLoader to passport-ui and use it in CardBody
