+++ apps/passport-client/components/screens/ProveScreen/SemaphoreSignatureProveScreen.tsx
@@ -18,6 +18,9 @@ import {
 import { cloneDeep } from "lodash";
 import { ReactNode, useCallback, useState } from "react";
 import styled from "styled-components";
+import { NewLoader } from "../../../../../packages/lib/passport-ui/src/NewLoader";
+import { Button2 } from "../../../new-components/shared/Button";
+import { Typography } from "../../../new-components/shared/Typography";
 import { appConfig } from "../../../src/appConfig";
 import { useIdentityV3, useSelf } from "../../../src/appHooks";
 import {
@@ -25,9 +28,6 @@ import {
   safeRedirectPending
 } from "../../../src/passportRequest";
 import { getHost, getOrigin, nextFrame } from "../../../src/util";
-import { Typography } from "../../../new-components/shared/Typography";
-import { Button2 } from "../../../new-components/shared/Button";
-import { NewLoader } from "../../../new-components/shared/NewLoader";
 
 export function SemaphoreSignatureProveScreen({
   req
