+++ apps/passport-client/components/screens/ZappScreens/ConnectPopupScreen.tsx
@@ -3,9 +3,9 @@ import { ReactNode, useCallback, useEffect, useRef, useState } from "react";
 import styled from "styled-components";
 import urljoin from "url-join";
 import * as v from "valibot";
+import { NewLoader } from "../../../../../packages/lib/passport-ui/src/NewLoader";
 import { BottomModalHeader } from "../../../new-components/shared/BottomModal";
 import { Button2 } from "../../../new-components/shared/Button";
-import { NewLoader } from "../../../new-components/shared/NewLoader";
 import { Typography } from "../../../new-components/shared/Typography";
 import { appConfig } from "../../../src/appConfig";
 import { useDispatch } from "../../../src/appHooks";
