+++ apps/passport-client/new-components/screens/Login/NewLoginInterstitialScreen.tsx
@@ -2,6 +2,7 @@ import { assertUnreachable } from "@pcd/util";
 import { useLayoutEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
+import { NewLoader } from "../../../../../packages/lib/passport-ui/src/NewLoader";
 import { AppContainer } from "../../../components/shared/AppContainer";
 import { useLoadedIssuedPCDs } from "../../../src/appHooks";
 import {
@@ -9,7 +10,6 @@ import {
   getPendingRequest
 } from "../../../src/sessionStorage";
 import { useSyncE2EEStorage } from "../../../src/useSyncE2EEStorage";
-import { NewLoader } from "../../shared/NewLoader";
 import { Typography } from "../../shared/Typography";
 
 export function NewLoginInterstitialScreen(): JSX.Element {
