+++ apps/passport-client/new-components/shared/Login/NewPasswordForm2.tsx
@@ -1,9 +1,9 @@
 import { Dispatch, ReactNode, SetStateAction, useRef, useState } from "react";
 import styled from "styled-components";
+import { NewLoader } from "../../../../../packages/lib/passport-ui/src/NewLoader";
 import { checkPasswordStrength } from "../../../src/checkPasswordStrength";
 import { PASSWORD_MINIMUM_LENGTH } from "../../../src/password";
 import { Button2 } from "../Button";
-import { NewLoader } from "../NewLoader";
 import { Typography } from "../Typography";
 import { PasswordInput2 } from "./PasswordInput2";
