+++ apps/passport-client/new-components/shared/Modals/DeleteAccountModal.tsx
@@ -1,10 +1,10 @@
 import { ReactElement, useCallback, useState } from "react";
 import styled from "styled-components";
+import { NewLoader } from "../../../../../packages/lib/passport-ui/src/NewLoader";
 import { useBottomModal, useDispatch, useSelf } from "../../../src/appHooks";
 import { BottomModal, BottomModalHeader } from "../BottomModal";
 import { Button2 } from "../Button";
 import { Input2 } from "../Input";
-import { NewLoader } from "../NewLoader";
 import { Typography } from "../Typography";
 
 export const DeleteAccountModal = (): ReactElement => {
