+++ packages/lib/passport-ui/src/index.ts
@@ -1,12 +1,13 @@
 export * from "./Button";
 export * from "./CopyIcon";
 export * from "./Core";
-export * from "./error";
 export * from "./HiddenText";
 export * from "./ImageZoom";
+export * from "./NewLoader";
 export * from "./QR";
 export * from "./SlidingTabs";
 export { css, keyframes, default as styled } from "./StyledWrapper";
 export type { FlattenSimpleInterpolation } from "./StyledWrapper";
 export * from "./Toggle";
 export * from "./VIcon";
+export * from "./error";
